refactor: rename EditorActions to EditorAction, use singular form
