+++ frontend/src/views/SqlEditor/EditorPanel/EditorAction.vue
